+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - User can manually specify llm system message instructions
+- File size limit of 10 MB implemented to avoid incorrect use (Only text files are accepted for now)
 
 ### Changed